feat: implement multi-step venue registration process with form state management
fix: update AccountSetup, PaymentDetails, VenueDetails, and VenueLocation components to accept props for data and state management
