+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "pubchat",
+  "name": "netwok_admin",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -25,6 +25,7 @@
     "@radix-ui/react-tabs": "^1.1.2",
     "@radix-ui/react-toast": "^1.2.11",
     "@radix-ui/react-tooltip": "^1.1.8",
+    "@tabler/icons-react": "^3.31.0",
     "@tailwindcss/postcss": "^4.0.11",
     "axios": "^1.7.9",
     "class-variance-authority": "^0.7.1",
@@ -32,7 +33,7 @@
     "cookie": "^1.0.2",
     "date-fns": "^4.1.0",
     "lucide-react": "^0.469.0",
-    "motion": "^12.6.2",
+    "motion": "^12.9.2",
     "nanoid": "^5.0.9",
     "next": "^15.2.3",
     "next-themes": "^0.4.6",
